2021.04/0430_pjt/pjt/requirements.txt
asgiref==3.3.4
Django==3.2
django-extensions==3.1.3
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
